Cell is now a plain UInt8

+++ Game Of Life - App/Classes/GridView.swift
@@ -192,7 +192,7 @@ class GridView: NSView
         }
     }
     
-    public func colorForAge( _ age: UInt64 ) -> NSColor
+    public func colorForAge( _ age: UInt8 ) -> NSColor
     {
         if( Preferences.shared.colors == false )
         {
